+++ examples/consumergroup/main.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"syscall"
 
-	"github.com/IBM/sarama"
+	"github.com/murugan100/sarama"
 )
 
 // Sarama configuration options