+++ mocks/sync_producer.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"sync"
 
-	"github.com/IBM/sarama"
+	"github.com/murugan100/sarama"
 )
 
 // SyncProducer implements sarama's SyncProducer interface for testing purposes.