+++ tools/kafka-console-producer/kafka-console-producer.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/rcrowley/go-metrics"
 
-	"github.com/IBM/sarama"
-	"github.com/IBM/sarama/tools/tls"
+	"github.com/murugan100/sarama"
+	"github.com/murugan100/sarama/tools/tls"
 )
 
 var (