+++ tools/kafka-producer-performance/main.go
@@ -15,8 +15,8 @@ import (
 
 	"github.com/rcrowley/go-metrics"
 
-	"github.com/IBM/sarama"
-	"github.com/IBM/sarama/tools/tls"
+	"github.com/murugan100/sarama"
+	"github.com/murugan100/sarama/tools/tls"
 )
 
 var (